### 2021-10-26 #3856 @whedon
[PRE REVIEW]: Phonemizer: Text to Phones Transcription for Multiple Languages in Python
**Submitting author:** @mmmaat (<a href="http://orcid.org/0000-0001-7586-7133">Bernard</a>)
**Repository:** <a href="https://github.com/bootphon/phonemizer" target ="_blank">https://github.com/bootphon/phonemizer</a>
**Version:** v3.0
