### 2021-11-29 #3856 @danielskatz
👋 @mmmaat - the next step is to find at least two reviewers. We have one already, and need at least one more. If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
