### 2021-11-29 #3856 @mmmaat
Hi @danielskatz, thank you for editing this paper. After a brief search in the reviewers list I fount 6 potential reviewers with expertise in Python and Natural Language Processing (more are available):
* sergioburdisso (Python, NLP)
* chrisbrickhouse (Python, linguistics)
* pzelasko (Python, speech, NLP)
* kmichael08 (Python, NLP)
* proycon (Python, NLP)
* aozorahime (Python, NLP)
