### 2021-11-29 #3856 @danielskatz
👋 @sergioburdisso & @chrisbrickhouse - would one or both of you be willing to review this JOSS submission?
