### 2021-11-29 bootphon/phonemizer#91 @chrisbrickhouse
State of the field and force alignment literature
*Part of a review at openjournals/joss-reviews#3958*

- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

The manuscript identifies a number of related software packages with which this program could interface, but there is a gap in its references to the force alignment literature. Within linguistics, especially phonetic analysis, force alignment is an important part of the research pipeline whereby an acoustic signal is segmented and aligned with a text transcript. This then allows corpus queries and phonetic analysis of segments. The paper briefly touched on this when discussing Kaldi (Povey, et al. 2011), but the state of the field is broader and this program has important implications for that field. I believe the paper would be improved by further review of that literature.

The most impactful piece of software in that field is the force Alignment and Vowel Extraction (FAVE) toolkit ([Rosenfelder, et al. 2014](https://doi.org/10.5281/zenodo.9846)) which converts orthographic transcriptions to phonetic transcriptions through dictionary lookups using the CMU pronunciation dictionary. This has the downside of not being able to handle out-of-dictionary words requiring experimenter transcription or data exclusion.

Other researchers have been trying to improve coverage of force alignment to underdocumented languages and a major problem is the lack of grapheme-to-phoneme mappings ([Barth, et al. 2020](https://repository.upenn.edu/pwpl/vol25/iss2/2 )) or comprehensive pronunciation dictionaries ([Johnson, Di Paolo, and Bell 2018](https://scholarspace.manoa.hawaii.edu/bitstream/10125/24763/1/johnson_et_al.pdf)). These can be substantial work and the language, orthographic system, or researcher time can limit the utility of these approaches.

These programs require a task similar to the one performed by this package, but do it in a seemingly different way. Comparing this package to the methods used in those packages will improve the paper by connecting it to a wider body of literature and identifying new potential areas of impact.

This is a really interesting project, and I'm excited to look further into the code!
