### 2021-11-29 bootphon/phonemizer#91 @mmmaat
Hi Christian, thanks for your comment. Indeed the references you are talking about seems interesting and -- indeed -- related to the phonemizer software. Just... I didn't know them! So I'll review this field and edit the paper in consequence.
