### 2021-11-30 bootphon/phonemizer#92 @henrykironde
Joss review 
@mmmaat, I think it is a great tool and a well-packaged software.

**These are some recommendations**(Not required):
- The docs(README) could use some visual improvements by wrapping all the command blocks(with 2 lines or more probably) in ```shell or ```j like the python code block. This may improve readability
- Include a code of conduct. Example: [Joss COC](https://github.com/openjournals/joss/blob/master/CODE_OF_CONDUCT.md), from the [covenant.org](https://www.contributor-covenant.org/)

- LICENSE.txt, I think It would be fine without an extension. (LICENSE)
- Could employ some config files to force styles during testing
`.style.yapf` by yapf   . i,e yapf --recursive path-to-source  --style=.style.yapf
`.pylintrc` and or `.flake8`
- Can you also explain a little on the pipeline given all the underlying backends as they go from text to grapheme, and then to phoneme.

**Qns**
What would you respond if asked about the performance of the tool?
Are there some limitations other than the number of supported languages or some edge cases like encoding, conflict resolutions?

**Requested changes:**

- [x] Add developer docs or instructions sections. This will help contributors with some of the policies, workflow, or getting started with environment setup, testing environments
- [x] I do not see the installation explanation for Windows and Mac users. https://github.com/bootphon/phonemizer#dependencies.
- [x] Include at least two ready to run Python usage examples in the docs(README)

ref: openjournals/joss-reviews#3958
