### 2021-12-03 9a55c22
Improved README
* Added a link to contribution guide for developers
* Added exemples in Python
* Added installation instruction for MacOS and Windows
