+++ README.md
@@ -1,7 +1,8 @@
 **Tests**
 [![Linux][badge-test-linux]](https://github.com/bootphon/phonemizer/actions/workflows/linux.yaml)
 [![MacOS][badge-test-macos]](https://github.com/bootphon/phonemizer/actions/workflows/macos.yaml)
-[![Windows][badge-test-windows]](https://github.com/bootphon/phonemizer/actions/workflows/windows.yaml) [![Codecov][badge-codecov]](https://codecov.io/gh/bootphon/phonemizer)
+[![Windows][badge-test-windows]](https://github.com/bootphon/phonemizer/actions/workflows/windows.yaml)
+[![Codecov][badge-codecov]](https://codecov.io/gh/bootphon/phonemizer)
 
 **Release**
 [![GitHub release (latest SemVer)][badge-github-version]](https://github.com/bootphon/phonemizer/releases/latest)
@@ -41,42 +42,84 @@
 
 ## Installation
 
-**You need python>=3.6.** If you really need to use python2, use an [older
-version](https://github.com/bootphon/phonemizer/releases/tag/v1.0) of
-the phonemizer.
-
-
 ### Dependencies
 
-* You need to install
-  [festival](http://www.festvox.org/docs/manual-2.4.0/festival_6.html#Installation),
-  [espeak-ng](https://github.com/espeak-ng/espeak-ng#espeak-ng-text-to-speech)
-  and [mbrola](https://github.com/numediart/MBROLA) on your system. On
-  Debian/Ubuntu simply run:
+**You need python>=3.6.** If you really need to use python2, use the
+[phonemizer-1.0] release.
 
-        $ sudo apt-get install festival espeak-ng mbrola
+You need to install
+[festival](http://www.festvox.org/docs/manual-2.4.0/festival_6.html#Installation),
+[espeak-ng](https://github.com/espeak-ng/espeak-ng#espeak-ng-text-to-speech)
+and/or [mbrola](https://github.com/numediart/MBROLA) in order to use the
+corresponding `phonemizer` backends. Follow instructions for your system below.
 
-* When using the **espeak-mbrola** backend, additional mbrola voices must be
-  installed (see
-  [here](https://github.com/espeak-ng/espeak-ng/blob/master/docs/mbrola.md)). On
-  Debian/Ubuntu, list the possible voices with `apt search mbrola`.
+<details><summary>on Debian/Unbuntu</summary>
+<p>
+
+To install dependencies, simply run `sudo apt-get install festival espeak-ng mbrola`.
+
+When using the **espeak-mbrola** backend, additional mbrola voices must be
+installed (see
+[here](https://github.com/espeak-ng/espeak-ng/blob/master/docs/mbrola.md)). List
+the installable voices with `apt search mbrola`.
+
+</p>
+</details>
+
+<details><summary>on CentOS/Fedora</summary>
+<p>
+
+To install dependencies, simply run `sudo yum install festival espeak-ng`.
+
+When using the **espeak-mbrola** backend, the mbrola binary and additional
+mbrola voices must be installed (see
+[here](https://github.com/espeak-ng/espeak-ng/blob/master/docs/mbrola.md)).
+
+</p>
+</details>
+
+<details><summary>on MacOS</summary>
+<p>
+
+**espeak** is available on brew at version 1.48: `brew install espeak`. If you
+want a more recent version you have to [compile it from
+sources](https://github.com/espeak-ng/espeak-ng/blob/master/docs/building.md#linux-mac-bsd).
+To install **festival***, **mbrola** and additional mbrola voices, use the
+script provided [here](https://github.com/pettarin/setup-festival-mbrola).
+
+</p>
+</details>
+
[…]
