### 2021-12-03 979bdd2
added backend properties table in README
