+++ README.md
@@ -19,25 +19,40 @@
 * The phonemizer allows simple phonemization of words and texts in many languages.
 
 * Provides both the `phonemize` command-line tool and the Python function
-  `phonemizer.phonemize`.
+  `phonemizer.phonemize`. See [function documentation][phonemize-function].
 
-* It is using four backends: espeak, espeak-mbrola, festival and segments.
+* It is based on four backends: **espeak**, **espeak-mbrola**, **festival** and
+  **segments**. The backends have different properties and capabilities resumed
+  in table below. The backend choice is let to the user.
 
-  * [espeak-ng](https://github.com/espeak-ng/espeak-ng) supports a lot of
-    languages and IPA (International Phonetic Alphabet) output.
+  * [espeak-ng](https://github.com/espeak-ng/espeak-ng) is a Text-to-Speech
+    software supporting a lot of languages and IPA (International Phonetic
+    Alphabet) output.
 
   * [espeak-ng-mbrola](https://github.com/espeak-ng/espeak-ng/blob/master/docs/mbrola.md)
     uses the SAMPA phonetic alphabet instead of IPA but does not preserve word
     boundaries.
 
-  * [festival](http://www.cstr.ed.ac.uk/projects/festival) currently supports
-    only American English. It uses a [custom
-    phoneset](http://www.festvox.org/bsv/c4711.html), but it allows tokenization
-    at the syllable level.
+  * [festival](http://www.cstr.ed.ac.uk/projects/festival) is another
+    Tex-to-Speech engine. Its phonemizer backend currently supports only
+    American English. It uses a [custom phoneset][festival-phoneset], but it
+    allows tokenization at the syllable level.
 
   * [segments](https://github.com/cldf/segments) is a Unicode tokenizer that
     build a phonemization from a grapheme to phoneme mapping provided as a file
     by the user.
+
+  |                              | espeak                   | espeak-mbrola           | festival                    | segments           |
+  | ---:                         | ---                      | ---                     | ---                         | ---                |
+  | **phone set**                | [IPA]                    | [SAMPA]                 | [custom][festival-phoneset] | user defined       |
+  | **supported languages**      | [100+][espeak-languages] | [10+][mbrola-languages] | US English                  | user defined       |
+  | **processing speed**         | fast                     | slow                    | very slow                   | fast               |
+  | **phone tokens**             | :heavy_check_mark:       | :heavy_check_mark:      | :heavy_check_mark:          | :heavy_check_mark: |
+  | **syllable tokens**          | :x:                      | :x:                     | :heavy_check_mark:          | :x:                |
+  | **word tokens**              | :heavy_check_mark:       | :x:                     | :heavy_check_mark:          | :heavy_check_mark: |
+  | **punctuation preservation** | :heavy_check_mark:       | :x:                     | :heavy_check_mark:          | :heavy_check_mark: |
+  | **stressed phones**          | :heavy_check_mark:       | :x:                     | :x:                         | :x:                |
+  | [**tie**][tie-IPA]           | :heavy_check_mark:       | :x:                     | :x:                         | :x:                |
 
 
 ## Installation
@@ -161,9 +176,7 @@ wish to contribute.
 ## Python usage
 
 In Python import the `phonemize` function with `from phonemizer import
-phonemize`. See
-[here](https://github.com/bootphon/phonemizer/blob/master/phonemizer/phonemize.py#L33)
[…]
