### 2021-12-03 bootphon/phonemizer#92 @mmmaat
Hi Henry,

Thank you very much for taking time to review phonemizer, for your comments and recommendations. Let me answer point by point. The code version I refeer to is on the [joss branch](https://github.com/bootphon/phonemizer/tree/joss).

**Recommendations**

* I added the pypi badge, and the JOSS one too.
* I added the `shell` style to command-line exemples, the rendering is a little bit clearer.
* I included a CODE_OF_CONDUCT.md and removed the extension on the LICENSE file
* About your remark on pipeline and backends explanation, I added a table in the README introduction explaining backends differences and capabilities.
* Except if you insist, I prefer not to include any `.style.yapf` because I never use any of that. Nevertheless I use some pylint or test coverage flags directly in the code when necessary, e.g. [here](https://github.com/bootphon/phonemizer/blob/joss/phonemizer/phonemize.py#L33) or [here](https://github.com/bootphon/phonemizer/blob/joss/phonemizer/phonemize.py#L158). The code is rated 9.9/10 by pylint.

**About the performance of the tool?**

In term of processing speed it depends on which backend is used. Segments is pure Python and reasonably fast. Espeak backend is very fast as it is binded directly to the C library using `ctypes` (before `phonemizer-3.0` it was based on `subprocess`, calling the espeak binary for each sentence, it was about 30 to 40 times slower than now...). The espeak-mbrola backend is a bit slower because it needs to go through speech wave generation (which is thrown away behind the scene). Finally festival is very slow because it is based on `subprocess`.

About the `phonemizer` implementation by itself, performance was not an objective during development. I was focused on stability and usability. For instance the punctuation preservation algorithm could be optimized. But this is not a critical section of the pipeline.

**Are there some limitations?**

The main limitation I see it that phonemizer depends on the bakend quality. For instance, espeak (which is by far the most used backend) introduces some important drawbacks: it can switch from one language to another during the phonemization of a single sentence, or can merge two words together, or even drop away some words. At the `phonemizer` level there is nothing we can do except to detect those mistakes and warn the user. Actually most of the issues I get on github are actually issues with espeak, not phonemizer.

**Requested changes**

* I added a [Developers](https://github.com/bootphon/phonemizer/tree/joss#developers) section on the README, redirecting to CONTRIBUTING.md. I do not go deep in details in this file, as I assume developers familiar with Python will know how to setup a venv or run unit tests. Please let me know if you think I should give more information.
* I added installation instruction for Windows and MacOS [here](https://github.com/bootphon/phonemizer/tree/joss#dependencies).
* I added two [exemples in Python](https://github.com/bootphon/phonemizer/tree/joss#python-usage).

Best,
Mathieu
