### 2021-12-04 bootphon/phonemizer#92 @henrykironde
Closing since all issues are covered. Thanks @mmmaat
