### 2021-12-04 bootphon/phonemizer#92 @henrykironde
@mmmaat, I have finished the review, all instructions work fine and easy to follow. All my test cases/scenarios worked fine as expected 👍🏿. One thing that you may want to consider is building the API docs for the whole source in the future but I understand the challenges.
