### 2021-12-04 bootphon/phonemizer#92 @mmmaat
Thanks! I know it would be a serious improvement to have a dedicated readthedocs page, it is planned as a middle-term objective.
