### 2021-12-04 19ab41d
minor corrections to README
