+++ paper/paper.bib
@@ -128,3 +128,20 @@
   doi={10.21437/Interspeech.2018-1456},
   url={http://dx.doi.org/10.21437/Interspeech.2018-1456}
 }
+
+@software{rosenfelder:2014,
+  author={Ingrid Rosenfelder and Josef Fruehwald and Keelan Evanini and Scott Seyfarth and Kyle Gorman and Hilary Prichard and Jiahong Yuan},
+  title={FAVE 1.1.3},
+  month=may,
+  year=2014,
+  publisher={Zenodo},
+  doi={10.5281/zenodo.9846},
+  url={https://doi.org/10.5281/zenodo.9846}
+}
+
+@inproceedings{mcauliffe:2017,
+  author={Michael McAuliffe and Michaela Socolof and Sarah Mihuc and Michael Wagner and Morgan Sonderegger},
+  title={Montreal Forced Aligner: trainable text-speech alignment using Kaldi},
+  booktitle={Proceedings of the 18th Conference of the International Speech Communication Association},
+  year={2017}
+}
