+++ paper/paper.md
@@ -51,7 +51,7 @@ and uses a non standard phoneset for transcription, but this backend is the only
 one to meet the requirement of some applications by preserving syllable
 boundaries. The third `phonemizer` backend is Segments [@forkel:2019], a Python
 package providing Unicode Standard tokenization routines and orthography
-segmentation. It relies on a grapheme to phone mapping to generate the
+segmentation. It relies on a grapheme to phoneme mapping to generate the
 transcription. This backend is mostly usefull for low-resource languages, for
 which users with linguistic expertise can write their own mappings. Six
 languages are provided as exemples with `phonemizer`: Chintang, Cree, Inuktitut,
@@ -60,15 +60,23 @@ Japanese, Sesotho and Yucatec.
 Text to phones transcription is a critical need in different applications
 related to natural language and speech processing. So far, the `phonemizer`
 package has been used in the preprocessing pipeline of various deep learning
-text to speech systems [@mozilla:2021; @asideas:2021;
-@espnet:2018]. It has also been used as a preprocessing step in word
-segmentation studies regarding the role of speech prosody in segmentability
-[@ludusan:2017] and the psychology of child development
-[@cristia:2019;@bernard:2020]. A phonetic transcription generated by the package
-was used to evaluate a phone discrimination task for the Zero Speech Challenge
-2017 [@dunbar:2017]. Finally, the `phonemizer` is very suitable to prepare
-datasets for their use with the Kaldi speech recognition toolkit [@povey:2011],
-where a phonetic transcription of text is a requirement for various algorithms.
+text to speech systems [@mozilla:2021; @asideas:2021; @espnet:2018]. It has also
+been used as a preprocessing step in word segmentation studies regarding the
+role of speech prosody in segmentability [@ludusan:2017] and the psychology of
+child development [@cristia:2019;@bernard:2020]. A phonetic transcription
+generated by the package was used to evaluate a phone discrimination task for
+the Zero Speech Challenge 2017 [@dunbar:2017]. The `phonemizer` is also very
+suitable to prepare datasets for their use with the Kaldi speech recognition
+toolkit [@povey:2011], where a phonetic transcription of text is a requirement
+for various algorithms. Finally, the package can be used to generate forced
+alignements of speech corpora, an important part of the research pipeline
+whereby an acoustic signal is segmented and aligned with a text transcript. The
+most impactful softwares in this field [@rosenfelder:2014, mcauliffe:2017]
+require a pronunciation dictionary to transcribe words into phonemes. Such
+dictionaries, when available, can be non-exhaustive, thus requiring experimenter
+transcription or data exclusion. Replacing the dictionary by the use of the
+`phonemizer` can therefore improve the alignament quality for compatible
+languages.
 
 
 # Acknowledgements
