### 2021-12-05 bootphon/phonemizer#91 @mmmaat
See changes [here](https://github.com/bootphon/phonemizer/blob/771ca5ed53c7f1d9ddf17860165d413abf3d00b0/paper/paper.md?plain=1#L71-L79).

I added references to FAVE and the Montreal Forced Aligner to the paper, explaining that the use of phonemizer instead of pronunciation dictionary can improve the overall pipeline for compatible languages. Nevertheless I choose to not introduce the issue of under-resourced languages because, even with the phonemizer, you need the language to be supported.

Do you agree with those corrections or do you expect more details?

Thanks.
