### 2021-12-07 bootphon/phonemizer#91 @chrisbrickhouse
I still think you should consider covering how phonemizer can be used for under-resourced languages, but I'll get back to that later.

One thing that is worth mentioning is that MFA also includes functionality for [training G2P models](https://montreal-forced-aligner.readthedocs.io/en/latest/user_guide/workflows/g2p_train.html). This seems different from how the `segments` backend in this program handles G2P for languages not supported out-of-the-box. As I understand it, MFA's G2P training requires an existing dictionary with sufficient data to train the model accurately. By contrast, the `segments` backend only seems to require a file that maps graphemes to their pronunciation. I believe the paper would be improved by comparing how this backend differs from the existing MFA function and discussing what use cases would work better for one or the other.

If my understanding is correct, this is where I think Phonemizer shows promise for lesser documented languages. Compiling a dictionary large enough to train a model on is a large task when compared to compiling a grapheme to phoneme map. For example one of my colleagues, Kate Lindsey, worked with the Ende Language Project to develop an orthography and dictionary for the Papuan language, with one of the goals being transparency in grapheme to phoneme mappings ([Lindsey 2021](https://www.cambridge.org/core/journals/journal-of-the-international-phonetic-association/article/abs/ende/8EFEC79687F73B691A1D6C91ABE93E4B)). In cases like these, compiling a grapheme-to-phoneme map for the `segments` backend is easier and more reliable than compiling a dictionary and training a model using MFA.

If I misunderstood how that backend works or you still think this is beyond the scope of your need statement, I won't push you to include it. To me it seems like a problem that this program can help solve, though I understand that automated transcription of Papuan languages is a rather niche use case. I'm unfamiliar with the NLP side of the literature review, but from my work with field linguists, I can see this being a very important tool for them which is why I'd encourage engaging with the literature on lesser documented languages.
