### 2021-12-10 a2f22ee
adding application to low-resource languages (issue 91)
