+++ paper/paper.md
@@ -28,11 +28,11 @@ phonetic alphabet is an important requirement in various applications related to
 speech and language processing, for instance for text to speech systems.
 `Phonemizer` is a Python package addressing precisely this issue: it transcribes
 a text from its orthographic representation into a phonetic one. The package is
-user-friendly and exposes a single `phonemize` function, also available as a
-command-line interface. It supports about a hundred different languages and
-provides end-user functionalities such as punctuation preservation, phones
-accentuation, tokenization at phone/syllable/word levels, as well as parallel
-processing of large input texts.
+user-friendly and exposes a single high-level `phonemize` function, a lower
+lovel API, and is also available as a command-line interface. It supports about
+a hundred different languages and provides end-user functionalities such as
+punctuation preservation, phones accentuation, tokenization at
+phone/syllable/word levels, as well as parallel processing of large input texts.
 
 
 # Statement of Need
@@ -43,19 +43,19 @@ backends, wrapped in an homogeneous interface by the package. The default
 backend used by `phonemizer` is eSpeak [@espeak:2019], a text to speech software
 built on linguistic expertise and hand written transcription rules. It
 transcribes text into the International Phonetic Alphabet and supports more than
-a hundred languages. Using MBROLA voices [@mbrola:2019], available for 35
+a hundred languages. Using MBROLA voices [@mbrola:2019], available for about 35
 languages, the eSpeak backend transcribes text in the SAMPA computer readable
 phonetic alphabet. Festival [@festival:2014] is another text to speech software
 used as a backend for `phonemizer`. It is available for American English only,
 and uses a non standard phoneset for transcription, but this backend is the only
 one to meet the requirement of some applications by preserving syllable
-boundaries. The third `phonemizer` backend is Segments [@forkel:2019], a Python
-package providing Unicode Standard tokenization routines and orthography
-segmentation. It relies on a grapheme to phoneme mapping to generate the
-transcription. This backend is mostly useful for low-resource languages, for
-which users with linguistic expertise can write their own mappings. Six
-languages are provided as examples with `phonemizer`: Chintang, Cree, Inuktitut,
-Japanese, Sesotho and Yucatec.
+boundaries of transcribed texts. The third `phonemizer` backend is Segments
+[@forkel:2019], a Python package providing Unicode Standard tokenization
+routines and orthography segmentation. It relies on a grapheme to phoneme
+mapping to generate the transcription. This backend is mostly useful for
+low-resource languages, for which users with linguistic expertise can write
+their own mappings. Six languages are provided as examples with `phonemizer`:
+Chintang, Cree, Inuktitut, Japanese, Sesotho and Yucatec.
 
 Text to phones transcription is a critical need in different applications
 related to natural language and speech processing. So far, the `phonemizer`
@@ -68,15 +68,21 @@ generated by the package was used to evaluate a phone discrimination task for
 the Zero Speech Challenge 2017 [@dunbar:2017]. The `phonemizer` is also very
 suitable to prepare datasets for their use with the Kaldi speech recognition
 toolkit [@povey:2011], where a phonetic transcription of text is a requirement
-for various algorithms. Finally, the package can be used to generate forced
-alignments of speech corpora, an important part of the research pipeline
-whereby an acoustic signal is segmented and aligned with a text transcript. The
-most impactful software in this field [@rosenfelder:2014;@mcauliffe:2017]
-require a pronunciation dictionary to transcribe words into phonemes. Such
-dictionaries, when available, can be non-exhaustive, thus requiring experimenter
-transcription or data exclusion. Replacing the dictionary by the use of the
-`phonemizer` can therefore improve the overall pipeline and alignment quality
-for compatible languages.
[…]
