+++ paper/paper.bib
@@ -145,3 +145,32 @@
   booktitle={Proceedings of the 18th Conference of the International Speech Communication Association},
   year={2017}
 }
+
+@article{barth:2020,
+  title={Using forced alignment for sociophonetic research on a minority language},
+  author={Barth, Danielle and Grama, James and Gonzalez, Simon and Travis, Catherine},
+  journal={University of Pennsylvania Working Papers in Linguistics},
+  volume={25},
+  number={2},
+  pages={2},
+  year={2020}
+}
+
+@article{johnson:2018,
+  title={Forced alignment for understudied language varieties: Testing Prosodylab-Aligner with Tongan data},
+  author={Johnson, Lisa M and Di Paolo, Marianna and Bell, Adrian},
+  year={2018},
+  publisher={University of Hawaii Press}
+}
+
+
+@article{cristia:2020,
+  title={Infant-directed input and literacy effects on phonological processing: Non-word repetition scores among the Tsimane’},
+  author={Cristia, Alejandrina and Farabolini, Gianmatteo and Scaff, Camila and Havron, Naomi and Stieglitz, Jonathan},
+  journal={Plos one},
+  volume={15},
+  number={9},
+  pages={e0237702},
+  year={2020},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
