### 2021-12-10 bootphon/phonemizer#91 @mmmaat
OK I see your point about under-documented languages now. Indeed the compilation of G2P maps to be used with the segments backend is a promising application of that tool. I'm not competent in linguistics (more in TTS or speech modeling) so I didn't get your point at the first time. So... at the end of the *Statement of Need* section I added few words on that point, along with 2 of your references, and one from my lab on the phonetic analysis of a Tsimane corpus ([Cristia, 2020](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0237702)). Still... I don't want to go deep into details because the paper is not about methodology, or comparing the accuracy of one method over another, this is just about introducing the software and its functionalities. Actually I don't know which approach is the better for which use case, I'm just offering another tool to the community.

About the G2P training in MFA, it still requires a non-exhaustive lexicon as input and is used to bootstrap the lexicon with missing words. There are also deep learning models ([that one](https://github.com/cmusphinx/g2p-seq2seq) for instance) doing G2P. This is a bit different from the phonemizer. Actually trained models could be integrated as new backends.
