### 2021-12-10 bootphon/phonemizer#91 @mmmaat
Changes [here](https://github.com/bootphon/phonemizer/commit/a2f22ee42ddc3e7f9e9fbd053660ab409483a149).
