### 2021-12-13 db89aaf
BaseBackend.phonemize raises if input is a str and not a list of str
