### 2021-12-13 f3ad760
BaseBackend.phonemize raises if input is a str (test added)
