### 2021-12-17 #3958 @danielskatz
👋 @mmmaat - it looks like we are close to done, but I note [the open issue](https://github.com/bootphon/phonemizer/issues/91) shown above from @chrisbrickhouse that appears to be blocking him from checking off at least one item on his review.
