### 2021-12-17 bootphon/phonemizer#91 @chrisbrickhouse
Sorry for the delay! I missed the notification during holiday travel, but LGTM. There's no need to go deep into details, and even just making clear that it's another potential tool is helpful. The wonderful thing about peer review is that we all get to learn from each other. The changes you made look great, and thanks for the conversation and your patience!
