### 2021-12-17 #3958 @danielskatz
👋 @mmmaat - the next step will be me proof-reading the paper, and then asking you to archive the software with a release tag - I'll try to do this fairly soon.
