### 2021-12-18 #3958 @danielskatz
👋 @mmmaat - Please work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch joss` to check again, and the command `@whedon generate pdf from branch joss` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.

In addition, I'm now proof-reading the paper and bibliography to see if there are any other issues.
