### 2021-12-18 #3958 @danielskatz
👋 @mmmaat - reading the paper, I'm a little surprised the intro paragraph mentions phones, speech, and text, but not phonemes, particularly for a package called phonemizer. Would it make sense to also mention phonemes here?

In addition, I've suggested some minor changes for language and bib entry cases in https://github.com/bootphon/phonemizer/pull/101
