+++ paper/paper.bib
@@ -141,7 +141,7 @@
 
 @inproceedings{mcauliffe:2017,
   author={Michael McAuliffe and Michaela Socolof and Sarah Mihuc and Michael Wagner and Morgan Sonderegger},
-  title={Montreal Forced Aligner: trainable text-speech alignment using Kaldi},
+  title={Montreal Forced Aligner: trainable text-speech alignment using {K}aldi},
   booktitle={Proceedings of the 18th Conference of the International Speech Communication Association},
   year={2017}
 }
