### 2021-12-18 bootphon/phonemizer#91 @mmmaat
Don't worry for the delay, and thanks again for reviewing this paper. You put me in a direction I didn't expect and paved the way to new applications for that tool, that was interesting!
