### 2021-12-18 129ee37
minor changes in paper and bib (#101)
Updated paper.bib and paper.md
