+++ paper/paper.bib
@@ -37,7 +37,7 @@
 }
 
 @misc{festival:2014,
-  title={The Festival Speech Synthesis System},
+  title={The {F}estival Speech Synthesis System},
   author={Black, Alan W. and Clark, Rob and Richmond, Korin and Yamagishi, Junichi and Oura, Keiichiro and King, Simon},
   year={2014},
   version={2.4},
@@ -47,7 +47,7 @@
 
 @misc{forkel:2019,
   author={Robert Forkel and Steven Moran and Johann-Mattis List and Simon J Greenhill and Lucas Ashby and Kyle Gorman and Gereon Kaiping},
-  title={cldf/segments: Unicode Standard tokenization},
+  title={{c}ldf/segments: Unicode Standard tokenization},
   year=2019,
   publisher={Zenodo},
   version={v2.1.3},
@@ -104,7 +104,7 @@
 
 @inproceedings{povey:2011,
   author = {Povey, Daniel and Ghoshal, Arnab and Boulianne, Gilles and Burget, Lukas and Glembek, Ondrej and Goel, Nagendra and Hannemann, Mirko and Motlicek, Petr and Qian, Yanmin and Schwarz, Petr and Silovsky, Jan and Stemmer, Georg and Vesely, Karel},
-  title = {The Kaldi Speech Recognition Toolkit},
+  title = {The {K}aldi Speech Recognition Toolkit},
   booktitle = {IEEE 2011 Workshop on Automatic Speech Recognition and Understanding},
   year = {2011},
   publisher = {IEEE Signal Processing Society},
@@ -141,7 +141,7 @@
 
 @inproceedings{mcauliffe:2017,
   author={Michael McAuliffe and Michaela Socolof and Sarah Mihuc and Michael Wagner and Morgan Sonderegger},
-  title={Montreal Forced Aligner: trainable text-speech alignment using Kaldi},
+  title={Montreal Forced Aligner: trainable text-speech alignment using {K}aldi},
   booktitle={Proceedings of the 18th Conference of the International Speech Communication Association},
   year={2017}
 }
@@ -157,7 +157,7 @@
 }
 
 @article{johnson:2018,
-  title={Forced alignment for understudied language varieties: Testing Prosodylab-Aligner with Tongan data},
+  title={Forced alignment for understudied language varieties: Testing {P}rosodylab-{A}ligner with {T}ongan data},
   author={Johnson, Lisa M and Di Paolo, Marianna and Bell, Adrian},
   year={2018},
   publisher={University of Hawaii Press}
