### 2021-12-18 #3958 @mmmaat
@danielskatz thanks for editing this paper. I will add the missing DOIs today. Thanks for correcting the typos in the text, I just merged your PR.

About your question about phones and phonemes... Actually phone/phonetic is related to text and phoneme/phonemic is related to speech. I made a mistake when calling that tool the `phonemizer` 6 years ago, it would have better been called `phonetizer` but... this is too late.
