+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={264--278},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/s13428-019-01223-3}
 }
 
 @misc{zhang:2020,
@@ -15,7 +16,8 @@
   year={2020},
   publisher={GitHub},
   journal={GitHub repository},
-  url={https://github.com/jxzhanggg/nonparaSeq2seqVC_code}
+  url={https://github.com/jxzhanggg/nonparaSeq2seqVC_code},
+  doi={10.1109/taslp.2019.2960721}
 }
 
 @misc{mozilla:2021,
@@ -82,7 +84,8 @@
   volume={3},
   pages={13--22},
   year={2019},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi={10.1162/opmi_a_00022}
 }
 
 @inproceedings{dunbar:2017,
@@ -91,7 +94,8 @@
   booktitle={2017 IEEE Automatic Speech Recognition and Understanding Workshop (ASRU)},
   pages={323--330},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/asru.2017.8268953}
 }
 
 @inproceedings{ludusan:2017,
@@ -116,7 +120,8 @@
   author={Hayashi, Tomoki and Yamamoto, Ryuichi and Inoue, Katsuki and Yoshimura, Takenori and Watanabe, Shinji and Toda, Tomoki and Takeda, Kazuya and Zhang, Yu and Tan, Xu},
   pages={7654--7658},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icassp40776.2020.9053512}
 }
 
 @inproceedings{espnet:2018,
@@ -143,7 +148,8 @@
   author={Michael McAuliffe and Michaela Socolof and Sarah Mihuc and Michael Wagner and Morgan Sonderegger},
   title={Montreal Forced Aligner: trainable text-speech alignment using {K}aldi},
   booktitle={Proceedings of the 18th Conference of the International Speech Communication Association},
-  year={2017}
+  year={2017},
+  doi={10.21437/interspeech.2017-1386}
 }
 
 @article{barth:2020,
@@ -172,5 +178,6 @@
   number={9},
   pages={e0237702},
   year={2020},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0237702}
 }
