### 2021-12-18 #3958 @mmmaat
> :wave: @mmmaat - the next step will be me proof-reading the paper, and then asking you to archive the software with a release tag - I'll try to do this fairly soon.

If everything it OK for you I will release phonemizer-3.0.1 with the changes made in the [README.md](https://github.com/bootphon/phonemizer/compare/joss?expand=1#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5) during the review and the [two minor bugs](https://github.com/bootphon/phonemizer/blob/master/CHANGELOG.md#not-yet-released) recently fixed. I prefer not to merge the `paper/` folder in the `main` branch, what do you think?
