### 2021-12-18 #3958 @danielskatz
It's fine not to merge the paper into the main branch.

At this point could you:
