### 2021-12-18 1ea8c17
Improvements from JOSS reviews
