+++ README.md
@@ -1,80 +1,141 @@
-[![Linux](https://github.com/bootphon/phonemizer/actions/workflows/linux.yaml/badge.svg?branch=master)](
-https://github.com/bootphon/phonemizer/actions/workflows/linux.yaml)
-[![MacOS](https://github.com/bootphon/phonemizer/actions/workflows/macos.yaml/badge.svg?branch=master)](
-https://github.com/bootphon/phonemizer/actions/workflows/macos.yaml)
-[![Windows](https://github.com/bootphon/phonemizer/actions/workflows/windows.yaml/badge.svg?branch=master)](
-https://github.com/bootphon/phonemizer/actions/workflows/windows.yaml)
-[![Codecov](https://img.shields.io/codecov/c/github/bootphon/phonemizer)](
-https://codecov.io/gh/bootphon/phonemizer) [![GitHub release (latest
-SemVer)](https://img.shields.io/github/v/release/bootphon/phonemizer)](
-https://github.com/bootphon/phonemizer/releases/latest)
-[![DOI](https://zenodo.org/badge/56728069.svg)](
-https://doi.org/10.5281/zenodo.1045825)
+**Tests**
+[![Linux][badge-test-linux]](https://github.com/bootphon/phonemizer/actions/workflows/linux.yaml)
+[![MacOS][badge-test-macos]](https://github.com/bootphon/phonemizer/actions/workflows/macos.yaml)
+[![Windows][badge-test-windows]](https://github.com/bootphon/phonemizer/actions/workflows/windows.yaml)
+[![Codecov][badge-codecov]](https://codecov.io/gh/bootphon/phonemizer)
+
+**Release**
+[![GitHub release (latest SemVer)][badge-github-version]](https://github.com/bootphon/phonemizer/releases/latest)
+[![PyPI][badge-pypi-version]](https://pypi.python.org/pypi/phonemizer)
+
+**Citation**
+[![status][badge-joss]](https://joss.theoj.org/papers/08d1ffc14f233f56942f78f3742b266e)
+[![DOI][badge-zenodo]](https://doi.org/10.5281/zenodo.1045825)
+
+---
 
 # Phonemizer -- *foʊnmaɪzɚ*
 
 * The phonemizer allows simple phonemization of words and texts in many languages.
 
 * Provides both the `phonemize` command-line tool and the Python function
-  `phonemizer.phonemize`.
+  `phonemizer.phonemize`. See [function documentation][phonemize-function].
 
-* It is using four backends: espeak, espeak-mbrola, festival and segments.
+* It is based on four backends: **espeak**, **espeak-mbrola**, **festival** and
+  **segments**. The backends have different properties and capabilities resumed
+  in table below. The backend choice is let to the user.
 
-  * [espeak-ng](https://github.com/espeak-ng/espeak-ng) supports a lot of
-    languages and IPA (International Phonetic Alphabet) output.
+  * [espeak-ng](https://github.com/espeak-ng/espeak-ng) is a Text-to-Speech
+    software supporting a lot of languages and IPA (International Phonetic
+    Alphabet) output.
 
   * [espeak-ng-mbrola](https://github.com/espeak-ng/espeak-ng/blob/master/docs/mbrola.md)
     uses the SAMPA phonetic alphabet instead of IPA but does not preserve word
     boundaries.
 
-  * [festival](http://www.cstr.ed.ac.uk/projects/festival) currently supports
-    only American English. It uses a [custom
-    phoneset](http://www.festvox.org/bsv/c4711.html), but it allows tokenization
-    at the syllable level.
+  * [festival](http://www.cstr.ed.ac.uk/projects/festival) is another
+    Tex-to-Speech engine. Its phonemizer backend currently supports only
[…]
