### 2021-12-18 #3958 @mmmaat
Here we are:
* The DOI: https://zenodo.org/record/5791097
* The tagged release: https://github.com/bootphon/phonemizer/releases/tag/v3.0.1
